# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "cyprus-dating API",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/v1/profiles/me": {
      "patch": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 40
                  },
                  "city": {
                    "type": "string",
                    "enum": [
                      "Limassol",
                      "Nicosia",
                      "Larnaca",
                      "Paphos",
                      "Paralimni",
                      "Ayia Napa"
                    ]
                  },
                  "languagesSpoken": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "enum": [
                        "en",
                        "uk",
                        "ru",
                        "ro",
                        "bg"
                      ]
                    },
                    "minItems": 1
                  },
                  "bio": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 280
                  },
                  "photoIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "maxItems": 3
                  }
                },
                "required": [
                  "firstName",
                  "city",
                  "languagesSpoken",
                  "bio"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profileId": {
                      "type": "string"
                    },
                    "accountId": {
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "age": {
                      "type": "integer"
                    },
                    "city": {
                      "type": "string",
                      "enum": [
                        "Limassol",
                        "Nicosia",
                        "Larnaca",
                        "Paphos",
                        "Paralimni",
                        "Ayia Napa"
                      ]
                    },
                    "languagesSpoken": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "en",
                          "uk",
                          "ru",
                          "ro",
                          "bg"
                        ]
                      }
                    },
                    "bio": {
                      "type": "string"
                    },
                    "photos": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "photoId": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "photoId",
                          "url"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "profileId",
                    "accountId",
                    "firstName",
                    "age",
                    "city",
                    "languagesSpoken",
                    "bio",
                    "photos"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profileId": {
                      "type": "string"
                    },
                    "accountId": {
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "age": {
                      "type": "integer"
                    },
                    "city": {
                      "type": "string",
                      "enum": [
                        "Limassol",
                        "Nicosia",
                        "Larnaca",
                        "Paphos",
                        "Paralimni",
                        "Ayia Napa"
                      ]
                    },
                    "languagesSpoken": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "en",
                          "uk",
                          "ru",
                          "ro",
                          "bg"
                        ]
                      }
                    },
                    "bio": {
                      "type": "string"
                    },
                    "photos": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "photoId": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "photoId",
                          "url"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "profileId",
                    "accountId",
                    "firstName",
                    "age",
                    "city",
                    "languagesSpoken",
                    "bio",
                    "photos"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "get": {
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profileId": {
                      "type": "string"
                    },
                    "accountId": {
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "age": {
                      "type": "integer"
                    },
                    "city": {
                      "type": "string",
                      "enum": [
                        "Limassol",
                        "Nicosia",
                        "Larnaca",
                        "Paphos",
                        "Paralimni",
                        "Ayia Napa"
                      ]
                    },
                    "languagesSpoken": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "en",
                          "uk",
                          "ru",
                          "ro",
                          "bg"
                        ]
                      }
                    },
                    "bio": {
                      "type": "string"
                    },
                    "photos": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "photoId": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "photoId",
                          "url"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "profileId",
                    "accountId",
                    "firstName",
                    "age",
                    "city",
                    "languagesSpoken",
                    "bio",
                    "photos"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/profiles/{profileId}": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "profileId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "profileId": {
                      "type": "string"
                    },
                    "accountId": {
                      "type": "string"
                    },
                    "firstName": {
                      "type": "string"
                    },
                    "age": {
                      "type": "integer"
                    },
                    "city": {
                      "type": "string",
                      "enum": [
                        "Limassol",
                        "Nicosia",
                        "Larnaca",
                        "Paphos",
                        "Paralimni",
                        "Ayia Napa"
                      ]
                    },
                    "languagesSpoken": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "en",
                          "uk",
                          "ru",
                          "ro",
                          "bg"
                        ]
                      }
                    },
                    "bio": {
                      "type": "string"
                    },
                    "photos": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "photoId": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "photoId",
                          "url"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "profileId",
                    "accountId",
                    "firstName",
                    "age",
                    "city",
                    "languagesSpoken",
                    "bio",
                    "photos"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/v1/profiles/me/photos": {
      "post": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "contentType": {
                    "type": "string",
                    "minLength": 1
                  },
                  "data": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "contentType",
                  "data"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "photoId": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "photoId",
                    "url"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    },
                    "details": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "code",
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}